Modularize zdtm.py
The zdtm.py file has grown over time to more than 2500 lines.
It contains a number of classes, functions and variables all
combined into a single monolithic python module.

This patch moves test classes, helper functions, flavors and exceptions
into separate files that are imported in zdtm.py. This makes the code
easier to extend and maintain in the future.
